+++ TheSadRogue.Integration.Templates/template_code/Integration.Project.SFML.CSharp/CustomPlayerKeybindingsComponent.cs
@@ -12,7 +12,7 @@ namespace TheSadRogue.Integration.Templates.SFML
     /// CUSTOMIZATION: Components can also be attached to maps, so the code for calling TakeTurn on all entities could
     /// be moved to a map component as well so that it is more re-usable by code that doesn't pertain to movement.
     /// </remarks>
-    internal class CustomPlayerKeybindingsComponent : PlayerKeybindingsComponent
+    internal class CustomKeybindingsComponent : PlayerKeybindingsComponent
     {
         protected override void MotionHandler(Direction direction)
         {